requirements.txt
altgraph==0.17
aniso8601==1.2.0
asn1crypto==1.4.0
cffi==1.14.2
click==7.1.2
cmake==3.18.2.post1
cryptography==2.1.4
dlib==19.21.0
git+git://github.com/ageitgey/face_recognition#egg=face_recognition
Flask==1.1.1
Flask-Ask==0.9.7
future==0.18.2
idna==2.10
imutils==0.5.3
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
opencv-python==4.4.0.42
pefile==2019.4.18
Pillow==7.2.0
pycparser==2.20
pygame==1.9.6
pyinstaller==4.1.dev0
pyinstaller-hooks-contrib==2020.10
pyOpenSSL==17.0.0
pyserial==3.4
python-dateutil==2.8.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
scipy==1.5.2
six==1.11.0
sudo==1.0.0
Werkzeug==0.16.1
